% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fitness_selection_differential.R
\name{calculate_fitness_selection_differential}
\alias{calculate_fitness_selection_differential}
\title{Calculate the selection differential caused by differences in the relative fitness of individuals in the population}
\usage{
calculate_fitness_selection_differential(
  sd.population.resistance,
  relative.fitness
)
}
\arguments{
\item{sd.population.resistance}{= The standard deviation of the resistance intensity values in the mosquito population}

\item{relative.fitness}{= The relative fitness of the most resistant individuals in the population compared to the least resistant individuals in the population.}
}
\description{
Calculate the selection differential caused by differences in the relative fitness of individuals in the population
}
