% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_initial_insecticide_dose.R
\name{calculate_initial_insecticide_dose}
\alias{calculate_initial_insecticide_dose}
\title{Calculate the Relative Dose of the Initial Application}
\usage{
calculate_initial_insecticide_dose(applied.dose, recommended.dose)
}
\arguments{
\item{applied.dose}{= The efficacy of the applied insecticide dose.}

\item{recommended.dose}{= The recommended efficacy of insecticide dose. Should normally be set at 1.}
}
\description{
Calculate the Relative Dose of the Initial Application
}
