% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_insecticide_exposure.R
\name{calculate_insecticide_exposure}
\alias{calculate_insecticide_exposure}
\title{Calculate the exposure to the insecticide in the intervention site.}
\usage{
calculate_insecticide_exposure(
  female.insecticide.exposure,
  male.insecticide.exposure
)
}
\arguments{
\item{female.insecticide.exposure}{= The proportion of female mosquitoes exposed to the insecticide}

\item{male.insecticide.exposure}{= The proportion of male mosquitoes exposed to the insecticide as a proportion of the female}
}
\value{
The overall proportion of mosquitoes exposed.
}
\description{
Calculate the exposure to the insecticide in the intervention site.
}
