% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_insecticide_selection_differential.R
\name{calculate_insecticide_selection_differential}
\alias{calculate_insecticide_selection_differential}
\title{Calculate the selection differential as a result of insecticidal selection pressure.}
\usage{
calculate_insecticide_selection_differential(
  current.resistance.intensity,
  sd.population.resistance,
  half.population.bioassay.survival.resistance = 900,
  conversion.factor = 0.48,
  intercept = 0.15,
  female.insecticide.exposure,
  male.insecticide.exposure,
  current.insecticide.efficacy
)
}
\arguments{
\item{current.resistance.intensity}{= The current resistance intensity in the location where the insecticide selection is occurring.}

\item{sd.population.resistance}{= The standard deviation of the resistance intensity values in the mosquito population}

\item{half.population.bioassay.survival.resistance}{= The resistance intensity that would give 50\% survival in the bioassay.}

\item{conversion.factor}{= The regression coefficient to convert bioassay survival to field survival}

\item{intercept}{= The intercept from the linear regression to convert bioassay survival to field survival}

\item{female.insecticide.exposure}{= The proportion of female mosquitoes exposed to the insecticide}

\item{male.insecticide.exposure}{= The proportion of male mosquitoes exposed to the insecticide as a proportion of the female exposure.}

\item{current.insecticide.efficacy}{= The current efficacy of the insecticide in deployment.}
}
\value{
The selection differential as a result of insecticide selection pressure.
}
\description{
Calculate the selection differential as a result of insecticidal selection pressure.
}
