% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_resistance_intensity_exposed_survivors.R
\name{calculate_resistance_intensity_exposed_survivors}
\alias{calculate_resistance_intensity_exposed_survivors}
\title{}
\usage{
calculate_resistance_intensity_exposed_survivors(
  initial.applied.efficacy,
  generations.since.deployment,
  basal.decay.rate,
  rapid.decay.rate,
  cut.off.generations,
  current.resistance.intensity,
  half.population.bioassay.survival.resistance = 900,
  conversion.factor = 0.48,
  intercept = 0.15
)
}
\description{

}
