% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_selection_differential_insecticide_exposed.R
\name{calculate_selection_differential_insecticide_exposed}
\alias{calculate_selection_differential_insecticide_exposed}
\title{Calculate the selection differential between those exposed and not exposed to the insecticide.}
\usage{
calculate_selection_differential_insecticide_exposed(
  sd.population.resistance,
  field.survival
)
}
\arguments{
\item{sd.population.resistance}{= The standard deviation of the resistance intensity in the population.}

\item{field.survival}{= The survival probability of individuals encountering the insecticide.}
}
\description{
Calculate the selection differential between those exposed and not exposed to the insecticide.
}
