% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bioassay_survival_to_field.R
\name{convert_bioassay_survival_to_field}
\alias{convert_bioassay_survival_to_field}
\title{Convert bioassay survival into field based survival}
\usage{
convert_bioassay_survival_to_field(
  bioassay.survival,
  conversion.factor = 0.48,
  intercept = 0.15,
  current.insecticide.efficacy
)
}
\arguments{
\item{bioassay.survival}{= The bioassay survival at a specified resistance intensity}

\item{conversion.factor}{= A regression coefficient to convert bioassay survival to field survival}

\item{intercept}{= intercept for equation.}
}
\description{
A formula to convert the bioassay survival into field
survival. Parameter estimation was conducted using linear modelling. See
Parameter Estimates: "Conversion of Bioassay Survival to Field Survival" for
details. Briefly, linear modelling was conducted on paired hut trial and bioassay
results to estimate the relationship. Data obtained from Churcher et al 2018.
}
