% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bioassay_survival_to_resistance.R,
%   R/plot_resistance_to_bioassay_survival.R
\name{\%>\%}
\alias{\%>\%}
\alias{plot_bioassay_survival_to_resistance}
\alias{plot_resistance_to_bioassay_survival}
\title{Plot the relationship between survival and resistance calculated from bioassay survival}
\usage{
plot_bioassay_survival_to_resistance(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  estimate.precision = 0.01,
  sd.population.resistance = 10,
  number.bioassays = 1000,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000,
  minimum.bioassay.survival = 0,
  maximum.bioassay.survival = 1,
  divisions = 0.01
)

plot_resistance_to_bioassay_survival(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  number.bioassays = 1000,
  minimum.resistance = 0,
  maximum.resistance = 25000,
  sd.population.resistance = 10
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1.}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{This is calculated using the calculate_half_population_resistance function}

\item{estimate.precision}{How precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{sd.population.resistance}{How much variation in the population resistance.}

\item{minimum.resistance.value}{Recommend setting as 0.}

\item{maximum.resistance.value}{This will depend on the half survival scale, but 25000 would be a good start.}

\item{minimum.bioassay.survival}{As a proportion, should be set as 0}

\item{maximum.bioassay.survival}{As a proportion, should be set as 1}

\item{divisions}{The resolution of the final plotted graph}

\item{minimum.resistance}{Recommend setting as 0.}

\item{maximum.resistance}{This will depend on the half survival scale, but 10000 would be a good start.}

\item{bioassay.survival}{The survival that occured in the bioassay, as a proportion(values must be between 0 and 1). Where 1=all survived, 0=all died}

\item{nsim}{How many replications of the rnorm function are conducted. Recommended value is 1000.}
}
\value{
A ggplot of the relationship between resistance (x axis) and survival (y axis).

A ggplot of the relationship between resistance (x axis) and survival (y axis).
}
\description{
Plot the relationship between survival and resistance calculated from bioassay survival

Plot the relationship between resistance and survival calculated from insecticide resistance
}
