% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_current_insecticide_efficacy.R
\name{calculate_current_insecticide_efficacy}
\alias{calculate_current_insecticide_efficacy}
\title{Calculate the current efficacy of the insecticide in deployment.}
\usage{
calculate_current_insecticide_efficacy(
  initial.applied.efficacy,
  generations.since.deployment,
  basal.decay.rate,
  rapid.decay.rate,
  cut.off.generations
)
}
\description{

}
