% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_initial_insecticide_dose.R
\name{calculate_initial_insecticide_dose}
\alias{calculate_initial_insecticide_dose}
\title{Calculate the Relative Dose of the Initial Application}
\usage{
calculate_initial_insecticide_dose(applied.dose, recommended.dose)
}
\description{
Calculate the Relative Dose of the Initial Application
}
